moves benchmark utils to own file
